Validation of Full name and first name
Changes to the class setName with validations:
Full name must contain a space between first and second name.
If does contains a space creates an array with first and second name then arrange them to the proper variables.
If there's no space, system throws an error (hopefully) with the correct syntax.