Change on Methods and Main variables
Changed variables to collect only Strings instead of parse the values to the correct data type, this will help to handle errors in the class.

The class setter methods now receives a String input and validate, if valid parse the value to the correct data type so it can be manipulated correctly if it can't be converted throws a error to the user and set the value of the int/double to 0.