Final Push.
On the main:
Moved the setDiscount method to the constructor, this way there's only functionalities to read, create object and write on the main.
Removed the function to write error, since I was able to fix the issues found on error handling on the class.

Renamed the output file document.

Added comments to finalize the code.

On the Class:
Fixing error messages.

The IllegalArgumentException was causing the error to not be recorded correctly on the name. I replaced with the setError method I created to handle the error messages. With this I was able to write on the document the proper message for each input error.

Added comments.